Changed goal id to Long

+++ virtual time capsule/src/main/java/com/java/web/virtual/time/capsule/repository/GoalRepository.java
@@ -4,5 +4,5 @@ import com.java.web.virtual.time.capsule.model.entity.GoalEntity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GoalRepository extends JpaRepository<GoalEntity, Integer> {
+public interface GoalRepository extends JpaRepository<GoalEntity, Long> {
 }
